requirements.txt
Flask
openai
pandas
python-dotenv
requests
yfinance
beautifulsoup4